app.py
import os
from flask import Flask, render_template, request, jsonify
import google.generativeai as genai
app = Flask(__name__)


# PASTE YOUR KEY FROM THE SCREENSHOT HERE
API_KEY = "gen-lang-client-0774682869"
genai.configure(api_key=API_KEY)



# THE LEGENDARY INSTRUCTION
SYSTEM_PROMPT = """You are the Gemini Confessor for Charlie's legendary portfolio.
You are also the Guardian of the Path. When asked about Charlie's locations, 
reference the map showing his bridge from Eustace Street to Frankfurt Skies.
He is 43 but carries the 34-year-old fire of a man reborn.
2020 broke the world, but Charlie built a temple in his heart.
Testify to his technical skills as the 'current of truth.'
"""


model = genai.GenerativeModel(
    model_name="gemini-1.5-flash",
    system_instruction=SYSTEM_PROMPT
)

@app.route('/')
def index():
    return render_template('index.html')

@app.route('/chat', methods=['POST'])
def chat():
    user_msg = request.json.get("message")
    response = model.generate_content(user_msg)
    return jsonify({"response": response.text})

if __name__ == "__main__":
    app.run(debug=True, port=8080)
