requirements.txt
flask==3.0.0
google-generativeai